+++ lib/UI/views/splash.dart
@@ -15,15 +15,13 @@ class SplashView extends StatelessWidget {
 
     return SafeArea(
       child: Scaffold(
-        backgroundColor:
-            brightness == Brightness.light ? Colors.white : Colors.black,
+        backgroundColor: brightness == Brightness.light ? Colors.white : Colors.black,
         body: SplashWidget<ConfigModel>(
           duration: const Duration(seconds: 8),
           backgroundWorker: loadConfig,
           routeName: onboarding ? '/onboarding' : '/home',
-          delegate: storageService.getDisplaySplashScreen()
-              ? MyAnimationDelegate()
-              : null,
+          delegate:
+              storageService.getDisplaySplashScreen() ? MyAnimationDelegate() : null,
         ),
       ),
     );
@@ -145,7 +143,7 @@ class _MarioState extends State<Mario> with SingleTickerProviderStateMixin {
     ).animate(
       CurvedAnimation(
         parent: widget.controller,
-        curve: Interval(
+        curve: const Interval(
           MARIO_SCALE_OUT_START,
           MARIO_SCALE_OUT_END,
           curve: Curves.linear,
@@ -158,7 +156,7 @@ class _MarioState extends State<Mario> with SingleTickerProviderStateMixin {
     ).animate(
       CurvedAnimation(
         parent: widget.controller,
-        curve: Interval(
+        curve: const Interval(
           MARIO_SCALE_IN_START,
           MARIO_SCALE_IN_END,
           curve: Curves.linear,
@@ -267,7 +265,7 @@ class _LootBoxState extends State<LootBox> {
     ).animate(
       CurvedAnimation(
         parent: widget.controller,
-        curve: Interval(
+        curve: const Interval(
           LOADING_WIDTH_START,
           LOADING_WIDTH_END,
           curve: Curves.linear,
@@ -283,10 +281,9 @@ class _LootBoxState extends State<LootBox> {
     final double boxWidth = (widget.parentSize.width / 3) * 0.8;
     final double boxHeight = boxWidth * 168.0 / 232.0;
 
-    final double boxX =
-        (2.7) * (widget.parentSize.width / 3) - (boxWidth / 2.0);
-    final double boxY = _translation.value *
-        ((widget.parentSize.height * 0.6) - boxHeight / 2.0);
+    final double boxX = (2.7) * (widget.parentSize.width / 3) - (boxWidth / 2.0);
+    final double boxY =
+        _translation.value * ((widget.parentSize.height * 0.6) - boxHeight / 2.0);
 
     return Stack(
       children: <Widget>[
@@ -313,9 +310,7 @@ class _LootBoxState extends State<LootBox> {
               Padding(
                 padding: EdgeInsets.only(right: boxWidth * 0.3),
                 child: Opacity(
-                  opacity: widget.controller.value >= LOADING_OPACITY_START
-                      ? 1.0
-                      : 0.0,
+                  opacity: widget.controller.value >= LOADING_OPACITY_START ? 1.0 : 0.0,
                   child: Container(
                     width: _textWidth.value * 200,
                     child: Text(