+++ lib/UI/widgets/series_filter.dart
@@ -80,7 +80,7 @@ class _SeriesFilterWidgetState extends State<_SeriesFilterWidget> {
             );
           },
         ) ??
-        false;
+        Future<bool>.value(false);
   }
 
   @override
@@ -186,8 +186,7 @@ class _SeriesFilterWidgetState extends State<_SeriesFilterWidget> {
           child: Scrollbar(
             child: ListView.separated(
               itemCount: series.length,
-              separatorBuilder: (BuildContext context, int index) =>
-                  const Divider(),
+              separatorBuilder: (BuildContext context, int index) => const Divider(),
               itemBuilder: (BuildContext context, int i) {
                 return Consumer<SeriesFilterProvider>(
                   builder: (_, SeriesFilterProvider provider, Widget child) {
@@ -221,9 +220,8 @@ class _SeriesFilterWidgetState extends State<_SeriesFilterWidget> {
   }
 
   void _setAll() {
-    final List<String> serieIds = assetsService.config.serieList
-        .map<String>((SerieModel s) => s.lKey)
-        .toList();
+    final List<String> serieIds =
+        assetsService.config.serieList.map<String>((SerieModel s) => s.lKey).toList();
 
     setState(() {
       filteredSeriesID