+++ lib/util/theme.dart
@@ -21,9 +21,8 @@ class ActionBarThemeData {
 
   Color get selectedItemColor => darkBlueColor;
 
-  Color get selectedItemBackgroundColor => (data.brightness == Brightness.light)
-      ? Colors.blue[50]
-      : Colors.blue[100];
+  Color get selectedItemBackgroundColor =>
+      (data.brightness == Brightness.light) ? Colors.blue[50] : Colors.blue[100];
 }
 
 class ItemCardThemeData {
@@ -35,9 +34,8 @@ class ItemCardThemeData {
 
   Color get backgroundColor => Colors.blue[50];
 
-  Color get shadowColor => (data.brightness == Brightness.light)
-      ? Colors.grey[300]
-      : Colors.grey[600];
+  Color get shadowColor =>
+      (data.brightness == Brightness.light) ? Colors.grey[300] : Colors.grey[600];
 
   Color get saturationColor => Colors.blue[50];
 
@@ -70,7 +68,7 @@ ThemeData buildTheme(Brightness brightness) {
   if (brightness == Brightness.dark) {
     base = ThemeData.dark();
     base = base.copyWith(
-      colorScheme: ColorScheme.dark(
+      colorScheme: const ColorScheme.dark(
         primary: lightBlueColor,
         primaryVariant: darkBlueColor,
         secondary: lightBlueColor,
@@ -86,17 +84,18 @@ ThemeData buildTheme(Brightness brightness) {
       primaryIconTheme: const IconThemeData(
         color: lightBlueColor,
       ), // Buttons of showAboutDialog.
-      buttonTheme: ButtonThemeData(
+      buttonTheme: const ButtonThemeData(
         buttonColor: lightBlueColor,
       ),
       floatingActionButtonTheme: const FloatingActionButtonThemeData(
         foregroundColor: Colors.white,
       ),
+      cursorColor: lightBlueColor,
     );
   } else {
     base = ThemeData.light();
     base = base.copyWith(
-      colorScheme: ColorScheme.light(
+      colorScheme: const ColorScheme.light(
         primary: lightBlueColor,
         primaryVariant: darkBlueColor,
         secondary: lightBlueColor,
@@ -123,13 +122,14 @@ ThemeData buildTheme(Brightness brightness) {
       primaryIconTheme: const IconThemeData(
         color: lightBlueColor,
       ), // Buttons of showAboutDialog.
-      buttonTheme: ButtonThemeData(
+      buttonTheme: const ButtonThemeData(
         textTheme: ButtonTextTheme.primary,
         buttonColor: lightBlueColor,
       ),
       floatingActionButtonTheme: const FloatingActionButtonThemeData(
         foregroundColor: Colors.white,
       ),
+      cursorColor: lightBlueColor,
     );
   }
   return base;